bugfix to the previous bugfix :P

+++ core/GBMemory.cc
@@ -109,7 +109,7 @@ void GBMemory::write(u16 addr, u8 value, WatchpointControl watch)
 	}
 }
 
-void GBMemory::write16(u16 addr, u8 value, WatchpointControl watch)
+void GBMemory::write16(u16 addr, u16 value, WatchpointControl watch)
 {
 	write(addr, value & 0xFF, watch);
 	write(addr+1, value >> 8, watch);
